refactor: remove init_coords action, retain the redux type and behavior
