--- packages/dnd-core/src/actions/dragDrop/local/initCoords.ts
@@ -1,97 +1,0 @@
-import {
-	Action,
-	DragDropManager,
-	XYCoord,
-	InitCoordsOptions,
-	InitCoordsPayload,
-	HandlerRegistry,
-	DragDropMonitor,
-	Identifier,
-} from '../../../interfaces'
-declare var require: any
-const invariant = require('invariant')
-import { INIT_COORDS } from '../types'
-
-const ResetCoordinatesAction = {
-	type: INIT_COORDS,
-	payload: {
-		clientOffset: null,
-		sourceClientOffset: null,
-	},
-}
-
-export function initCoords<Context>(
-	manager: DragDropManager<Context>,
-	sourceIds: string[] = [],
-	{ clientOffset, getSourceClientOffset }: InitCoordsOptions,
-): Action<InitCoordsPayload> | undefined {
-	const monitor = manager.getMonitor()
-	const registry = manager.getRegistry()
-
-	// Initialize the client offset before canDrag is invoked
-	manager.dispatch({
-		type: INIT_COORDS,
-		payload: {
-			sourceClientOffset: null,
-			clientOffset: clientOffset || null,
-		},
-	})
-
-	verifyInvariants(sourceIds, monitor, registry)
-
-	const sourceId = getDragSourceId(sourceIds, monitor)
-	if (sourceId === null) {
-		// reset coordinates that might have been set in an "INIT_CLIENT_OFFSET" as no drag is happening
-		return ResetCoordinatesAction
-	}
-	const sourceClientOffset: XYCoord | null = determineSourceClientOffset(
-		sourceId,
-		clientOffset,
-		getSourceClientOffset,
-	)
-	return {
-		type: INIT_COORDS,
-		payload: {
-			clientOffset: clientOffset || null,
-			sourceClientOffset: sourceClientOffset || null,
-		},
-	}
-}
-
-function verifyInvariants(
-	sourceIds: string[],
-	monitor: DragDropMonitor,
-	registry: HandlerRegistry,
-) {
-	invariant(!monitor.isDragging(), 'Cannot call initCoords while dragging.')
-	for (const s of sourceIds) {
-		invariant(registry.getSource(s), 'Expected sourceIds to be registered.')
-	}
-}
-
-function getDragSourceId(sourceIds: string[], monitor: DragDropMonitor) {
-	let sourceId = null
-	for (let i = sourceIds.length - 1; i >= 0; i--) {
-		if (monitor.canDragSource(sourceIds[i])) {
-			sourceId = sourceIds[i]
-			break
-		}
-	}
-	return sourceId
-}
-
-function determineSourceClientOffset(
-	sourceId: string,
-	clientOffset: XYCoord | undefined,
-	getSourceClientOffset: ((sourceId: Identifier) => XYCoord) | undefined,
-) {
-	let result: XYCoord | null = null
-	if (clientOffset) {
-		invariant(
-			typeof getSourceClientOffset === 'function',
-			'When clientOffset is provided, getSourceClientOffset must be a function.',
-		)
-		result = getSourceClientOffset!(sourceId)
-	}
-	return result
-}
